edit/add/delete fully functional, conformed styles between tasks and notes
