+++ client/src/components/Tasks/Tasks.scss
@@ -4,7 +4,7 @@
 .todos {
 	text-align: right;
 	background: $app-background-dark;
-	padding: 1rem 1rem 0.5rem;
+	padding: 0.5rem 1rem;
 	border-radius: $border-raduis;
 
 	&__header {
